added tweetManagerService
